Xperf DPC-ISR Capture - V1
The current version of this script generates a folder sorted in ascending order, specifically from 'Test 1' to 'Test N.' It is essential that the 'Perfmon' directory in the designated path is devoid of remnants from previous tests; otherwise, the script will raise an error. Furthermore, each time the script is executed, it inspects the previous file, e.g., 'Test 1,' and creates, saving as 'Test 2,' an 'trace.etl' file. Simultaneously, it generates a text file exclusively dedicated to DPC & ISR data for a simplified and fundamental analysis. The text file, named 'iteration,' is purely auxiliary, providing a numerical value to keep track of the current test iteration.