Xperf DPC-ISR Capture - V2
The version of this script introduces a fundamental change: upon the initial execution, it will automatically proceed with subsequent tests, eliminating the need for manual script execution for each test. This modification aims to enhance convenience and ensure the avoidance of inaccurate results. It is important to note that any adjustments to the number of tests require manual intervention in the script. By default, the script is configured for 31 tests. The reason for this specific number is that the first test is excluded from counting to circumvent potential inaccuracies associated with the initial run. As a result, my analysis encompasses tests 2 through 31, constituting a total of 30 tests.